30706 refactor to reuse code
